Vector Math

+++ src/main.cpp
@@ -1,17 +1,18 @@
 #include <stdio.h>
+#include <cstdio>
 #include <stdlib.h>
 #include <stdbool.h>  // Include for bool type if needed
 #include <math.h>
 #include <thread>
 #include <algorithm> 
-#include <vectorMath.cpp>
+#include "../include/vectorMath.h"
 
 
 
 
 
 int main(){
-    int vec[3] = {1,1,1};
+    float vec[3] = {1,1,1};
     printf("f/n",vectorMag(vec));
     return 0;
 }
